refactor: api & extension
