+++ runtime/devnet/src/config/api.rs
@@ -1,8 +1,54 @@
-use crate::{
-	config::assets::TrustBackedAssetsInstance,
-	fungibles::{self},
-	Runtime,
-};
+use crate::{config::assets::TrustBackedAssetsInstance, fungibles, Runtime, RuntimeCall};
+use codec::{Decode, Encode, MaxEncodedLen};
+use frame_support::traits::Contains;
+
+/// A query of runtime state.
+#[derive(Encode, Decode, Debug, MaxEncodedLen)]
+#[repr(u8)]
+pub enum RuntimeRead {
+	/// Fungible token queries.
+	#[codec(index = 150)]
+	Fungibles(fungibles::Read<Runtime>),
+}
+
+/// A type to identify allowed calls to the Runtime from the API.
+pub struct AllowedApiCalls;
+
+impl Contains<RuntimeCall> for AllowedApiCalls {
+	/// Allowed runtime calls from the API.
+	fn contains(c: &RuntimeCall) -> bool {
+		use fungibles::Call::*;
+		matches!(
+			c,
+			RuntimeCall::Fungibles(
+				transfer { .. }
+					| transfer_from { .. }
+					| approve { .. } | increase_allowance { .. }
+					| decrease_allowance { .. }
+					| create { .. } | set_metadata { .. }
+					| start_destroy { .. }
+					| clear_metadata { .. }
+					| mint { .. } | burn { .. }
+			)
+		)
+	}
+}
+
+impl Contains<RuntimeRead> for AllowedApiCalls {
+	/// Allowed state queries from the API.
+	fn contains(c: &RuntimeRead) -> bool {
+		use fungibles::Read::*;
+		matches!(
+			c,
+			RuntimeRead::Fungibles(
+				TotalSupply(..)
+					| BalanceOf { .. } | Allowance { .. }
+					| TokenName(..) | TokenSymbol(..)
+					| TokenDecimals(..) | AssetExists(..)
+			)
+		)
+	}
+}
 
 impl fungibles::Config for Runtime {
 	type AssetsInstance = TrustBackedAssetsInstance;
